feat: Replace preview.ts with preview.tsx for enhanced Storybook configuration
- Deleted the old preview.ts file and introduced preview.tsx.
- Added a StoryWrapper component for consistent styling and tooltip context.
- Updated Storybook parameters and added decorators for improved component rendering.
